requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.35
alembic==1.16.2
pydantic==2.9.2
pydantic-settings==2.5.2
groq>=0.6.0
aiosmtpd>=1.4.5
apscheduler==3.10.4
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
structlog==25.5.0
python-dotenv==1.0.1
email-validator==2.2.0
python-multipart==0.0.12

